many: rename FindInPath to LookupPath
